stringvalue escaper
